fix foreach on null

+++ src/Controllers/StandsTypesController.php
@@ -190,19 +190,21 @@ class StandsTypesController extends Controller
 
             $old_stand_categories = $update_stand_type->categories()->pluck('category_id')->toArray();
 
-            // insert the new ones that are not in the old array
-            foreach($request->category_id as $index => $category_id) {
-                if(!in_array($category_id, $old_stand_categories)) {
-                    $update_stand_type->categories()->create([
-                        'category_id' => $category_id,
-                    ]);
+            if(!is_null($request->category_id)) {
+                // insert the new ones that are not in the old array
+                foreach($request->category_id as $index => $category_id) {
+                    if(!in_array($category_id, $old_stand_categories)) {
+                        $update_stand_type->categories()->create([
+                            'category_id' => $category_id,
+                        ]);
+                    }
                 }
-            }
 
-            // delete the old ones that are not in the new array
-            foreach($old_stand_categories as $index => $category_id) {
-                if(!in_array($category_id, $request->category_id)) {
-                    $update_stand_type->categories()->where('category_id', $category_id)->delete();
+                // delete the old ones that are not in the new array
+                foreach($old_stand_categories as $index => $category_id) {
+                    if(!in_array($category_id, $request->category_id)) {
+                        $update_stand_type->categories()->where('category_id', $category_id)->delete();
+                    }
                 }
             }
